requirements.txt
# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1

# Data Processing
pandas==2.1.3
numpy==1.24.3

# AI Frameworks
langchain==0.1.0
langchain-community
langchain-groq==0.0.1

# LLM APIs
groq==0.4.1
anthropic==0.8.1

# Databases
pymongo==4.6.1
chromadb==0.4.18

# UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

google-genai    

reportlab==4.0.7